generate fake data in testutils/common/data.common.testutil.js
and simplify testutils/common/data.common.testutil.js a bit
